+++ index.php
@@ -12,15 +12,17 @@ License: MIT
 namespace WebStudentRecordBook;
 
 use StudentUtility\API;
+use WebStudentRecordBook\Api\UserDataByStudController;
 use WebStudentRecordBook\Menu\AdminMenu;
 use WebStudentRecordBook\Widget\WebStudentRecordBookWidget;
 
 include_once(ABSPATH . 'wp-admin/includes/plugin.php');
 
 if (is_plugin_active('StudentUtility-wp-plugin/index.php')) {
-    require 'Api/RestController.php';
+    require 'Api/UserDataByStudController.php';
     require 'Widget/WebStudentRecordBookWidget.php';
     require 'Menu/AdminMenu.php';
     new WebStudentRecordBookWidget(API::getApiInstance());
     new AdminMenu();
+    new UserDataByStudController(API::getApiInstance()->getRepository());
 }
